+++ ChangeLog.md
@@ -10,5 +10,11 @@
   - Rename some class and methods might not matched with old version.
   - Add plugins check, if the renderer plugin is not installed, the library will not be loaded.
 
+- ### 1.0.1
+  - Chnage the host object of AOVHelper to VideoPost(from document)
+  - Add methods to modify octane light mask on octane tag
+  - Fix some hint
+  - Fix Octane Scene parameter
+  - Update the example of AOVHelper
 ---
 - __coming soon...__